+++ src/main/java/calculator/domain/Operation.java
@@ -1,16 +1,13 @@
-package calculator.util;
+package calculator.domain;
 
 import java.util.Arrays;
-import java.util.Map;
 import java.util.Optional;
 import java.util.function.DoubleBinaryOperator;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 public enum Operation {
-    PLUS  ("+", (x, y) -> x + y),
-    MINUS ("-", (x, y) -> x - y),
-    TIMES ("*", (x, y) -> x * y),
+    PLUS("+", (x, y) -> x + y),
+    MINUS("-", (x, y) -> x - y),
+    TIMES("*", (x, y) -> x * y),
     DIVIDE("/", (x, y) -> x / y);
 
     private final String symbol;
@@ -25,11 +22,6 @@ public enum Operation {
         return op.applyAsDouble(x, y);
     }
 
-    @Override
-    public String toString() {
-        return symbol;
-    }
-
     public static Optional<Operation> from(String symbol) {
         return Arrays.stream(values())
             .filter(operation -> operation.isSameSymbol(symbol))
@@ -38,5 +30,10 @@ public enum Operation {
 
     private boolean isSameSymbol(String symbol) {
         return this.symbol.equals(symbol);
+    }
+
+    @Override
+    public String toString() {
+        return symbol;
     }
 }
