+++ src/main/java/calculator/ui/FormulaInput.java
@@ -2,9 +2,6 @@ package calculator.ui;
 
 import calculator.domain.Formula;
 import calculator.util.Console;
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
 
 public class FormulaInput {
 
@@ -12,18 +9,16 @@ public class FormulaInput {
     }
 
     public static Formula getFormula() {
-        Formula formular = null;
         System.out.println("계산할 연산식을 입력해주세요.");
         try {
             String input = Console.readLine();
             isBlankOrNot(input);
 
-            formular = new Formula(input);
+            return new Formula(input);
         } catch (IllegalArgumentException e) {
             System.out.println("[ERROR] " + e.getMessage());
             return getFormula();
         }
-        return formular;
     }
 
     private static void isBlankOrNot(String input) {
